ixplore: Move defs and functs to header for reuse
